+++ controller/user_query.js
@@ -1,19 +1,22 @@
 const bcrypt = require("bcrypt");
 const { Query } = require("pg");
+const jwt = require("jsonwebtoken");
+require("dotenv").config();
+//validator implementation
+const validator = require("validator");
 
 const Pool = require("pg").Pool;
 const pool = new Pool({
-  user: "postgres",
-  host: "postgres",
-  database: "my_db",
-  password: "postgres", //find how to hide/encrypt the password
-  port: 5432,
+  user: `${process.env.PGUSERNAME}`,
+  database: `${process.env.DATABASE_URL}`,
+  password: `${process.env.PGPASSWORD}`,
+  port: process.env.PGPORT,
 });
 
-const searchUser = async function (first_name, last_name) {
+const searchUser = async function (user_id) {
   let query = {
-    text: "select * from quatro_user where first_name=$1 and last_name=$2",
-    values: [first_name, last_name],
+    text: "select email, first_name, last_name, date_of_birth, gender, user_credit from quatro_user where user_id = $1 ",
+    values: [user_id],
   };
 
   let resultQuery = await pool.query(query);
@@ -26,67 +29,135 @@ const searchUser = async function (first_name, last_name) {
 };
 
 const searchUserAPI = async (request, response) => {
-  const { first_name, last_name } = request.body;
-
   try {
-    let fl_name = await searchUser(first_name, last_name);
-    response.status(200).json({ result: fl_name });
+    let userDeets = await searchUser(request.query.user_id);
+    response.status(200).json({ result: userDeets });
   } catch (error) {
     response.status(404).json({ error: error.message });
   }
 };
 
+const createToken = (user_id) => {
+  return jwt.sign({ user_id: user_id }, process.env.SECRET, {
+    expiresIn: "1d",
+  });
+};
+
 const loginUser = async function (email, password) {
   let query = {
-    text: "select * from quatro_user where email=$1",
+    text: "select email, password, user_id from quatro_user where email=$1",
     values: [email],
   };
 
   let resultQuery = await pool.query(query);
   let user = resultQuery.rows;
 
+  //validation
+
   if (user.length === 0) {
     throw Error("Email doesnt exist");
   }
 
-  console.log(user);
+  //------------------------
+
   let validPassword = await bcrypt.compare(password, user[0]["password"]);
 
   if (!validPassword) {
-    throw Error("Invalid Password");
+    throw Error("Incorrect Password");
   }
-  return user[0];
+  return user[0].user_id;
 };
 
 const loginAPI = async (request, response) => {
   const { email, password } = request.body;
   try {
     let user = await loginUser(email, password);
-    response.status(200).json({ result: user });
+    let userJwt = createToken(user);
+    response.cookie("token", userJwt, {
+      httpOnly: true,
+      sameSite: "None",
+      secure: true,
+    });
+    response.status(200).json({ result: email, userJwt });
   } catch (error) {
     response.status(404).json({ error: error.message });
   }
 };
 
-const createUser = async function (email, password) {
+const createUser = async function (
+  email,
+  password,
+  first_name,
+  last_name,
+  date_of_birth,
+  gender
+) {
   let query_1 = {
-    text: "select * from quatro_user where email=$1",
+    text: "select email, password from quatro_user where email=$1",
     values: [email],
   };
 
   let resultQuery_1 = await pool.query(query_1);
   let user = resultQuery_1.rows;
+  var regName = /^[A-Za-z'\s]*$/;
+  // var regName= /^([a-z]+[,.]?[ ]?|[a-z]+['-]?)+$/;
 
   if (user.length !== 0) {
     throw Error("Email exist");
+  }
+
+  if (
+    first_name == "" ||
+    !regName.test(first_name) ||
+    last_name == "" ||
+    !regName.test(last_name)
+  ) {
+    throw Error("Name should contain alphabets only");
+  }
+
+  if (!gender) {
+    throw Error("Please select gender option");
+  }
+  if (!email && !password) {
+    throw Error("Email and password field cannot be empty");
+  }
+
+  if (!email || !email.trim()) {
+    throw Error("Email field cannot be empty");
+  }
+
+  if (!password || !password.trim()) {
+    throw Error("Password field cannot be empty");
+  }
+
+  if (!validator.isEmail(email)) {
+    throw Error("Email not valid");
+  }
+
+  if (password.length < 8) {
+    throw Error("Password should consists at least 8 characters");
+  }
+
+  if (
+    !validator.isStrongPassword(password, {
+      minLength: 8,
+      minLowercase: 1,
+      minUppercase: 1,
+      minNumbers: 1,
+      minSymbols: 1,
+    })
+  ) {
+    throw Error(
+      "Password should consists at least 1 lowercase, 1 uppercase, 1 number and 1 symbol "
+    );
   }
 
   const salt = await bcrypt.genSalt(10);
   const passHash = await bcrypt.hash(password, salt);
 
   let query = {
-    text: "insert into quatro_user(email,password,user_credit) values ($1,$2,100) returning user_id",
-    values: [email, passHash],
+    text: "insert into quatro_user(email,password,first_name,last_name,date_of_birth,gender,user_credit) values ($1,$2,$3,$4,$5,$6,100) returning user_id",
+    values: [email, passHash, first_name, last_name, date_of_birth, gender],
   };
 
   let resultQuery = await pool.query(query);
@@ -96,10 +167,19 @@ const createUser = async function (email, password) {
 };
 
 const createUserAPI = async (request, response) => {
-  const { email, password } = request.body;
+  const { email, password, first_name, last_name, date_of_birth, gender } =
+    request.body;
   try {
-    let newUser = await createUser(email, password);
-    response.status(200).json({ result: newUser });
+    let newUser = await createUser(
+      email,
+      password,
+      first_name,
+      last_name,
+      date_of_birth,
+      gender
+    );
+
+    response.status(200).json({ result: email, message: "User Created" });
   } catch (error) {
     console.log("error:", error);
     response.status(404).json({ error: error.message });
@@ -110,33 +190,39 @@ const updateUser = async function (
   first_name,
   last_name,
   date_of_birth,
-  phone_number,
   email,
-  password,
+  oldPassword,
   user_id
 ) {
-  const salt = await bcrypt.genSalt(10);
-  const passHash = await bcrypt.hash(password, salt);
-  if (isNaN(phone_number)) {
-    throw error("Invalid phone number");
+  let query_1 = {
+    text: "select email, password from quatro_user where user_id=$1",
+    values: [user_id],
+  };
+
+  let resultQuery_1 = await pool.query(query_1);
+  let user = resultQuery_1.rows;
+
+  if (user.length === 0) {
+    throw Error("User doesnt exist");
   }
+
+  let validPassword = false;
+
+  if (oldPassword) {
+    //empty
+    validPassword = await bcrypt.compare(oldPassword, user[0]["password"]);
+  }
+
+  if (!validPassword) {
+    throw Error("Invalid Password");
+  }
+
   let query = {
     text: `update quatro_user set first_name = coalesce(nullif($1,''), first_name),
            last_name = coalesce(nullif($2,''), last_name),
            date_of_birth = coalesce(nullif($3,''), date_of_birth),
-           phone_number = coalesce(nullif($4,''), phone_number),
-           email = coalesce(nullif($5,''), email),
-           password = coalesce(nullif($6,''), password)
-           where user_id = $7`,
-    values: [
-      first_name,
-      last_name,
-      date_of_birth,
-      phone_number,
-      email,
-      passHash,
-      user_id,
-    ],
+           email = coalesce(nullif($4,''), email) where user_id = $5`,
+    values: [first_name, last_name, date_of_birth, email, user_id],
   };
 
   let resultQuery = await pool.query(query);
@@ -146,28 +232,80 @@ const updateUser = async function (
 };
 
 const updateUserAPI = async (request, response) => {
-  const {
-    first_name,
-    last_name,
-    date_of_birth,
-    phone_number,
-    email,
-    password,
-    user_id,
-  } = request.body;
+  const { first_name, last_name, date_of_birth, email, oldPassword, user_id } =
+    request.body;
+
   try {
-    let updateUserDB = await updateUser(
+    await updateUser(
       first_name,
       last_name,
       date_of_birth,
-      phone_number,
       email,
-      password,
+      oldPassword,
       user_id
     );
-    response
-      .status(200)
-      .json({ result: updateUserDB, message: "User updated" });
+
+    //const updateUserJwt = createToken(updateUserDB?.user_id);
+
+    response.status(200).json({ result: email, message: "User updated" });
+  } catch (error) {
+    console.log("error:", error);
+    response.status(404).json({ error: error.message });
+  }
+};
+
+const updatePassword = async function (oldPassword, password, user_id) {
+  const salt = await bcrypt.genSalt(10);
+
+  console.log(`passws ${salt} ${password}`);
+  const passHash = await bcrypt.hash(password, salt);
+
+  // if (isNaN(phone_number)) {
+  //   throw error("Invalid phone number");
+  // }
+  let query_1 = {
+    text: "select email, password from quatro_user where user_id=$1",
+    values: [user_id],
+  };
+
+  let resultQuery_1 = await pool.query(query_1);
+  let user = resultQuery_1.rows;
+
+  if (user.length === 0) {
+    throw Error("User doesnt exist");
+  }
+
+  let validPassword = true;
+
+  if (oldPassword) {
+    //empty
+    validPassword = await bcrypt.compare(oldPassword, user[0]["password"]);
+  }
+
+  if (!validPassword) {
+    throw Error("Invalid Password");
+  }
+
+  let query = {
+    text: `update quatro_user set password = coalesce(nullif($1,''), password) where user_id = $2`,
+    values: [passHash, user_id],
+  };
+
+  let resultQuery = await pool.query(query);
+  let passwordUpdate = resultQuery.rows;
+
+  return passwordUpdate[0];
+};
+
+const updatePasswordAPI = async (request, response) => {
+  const { oldPassword, password, user_id } = request.body;
+
+  try {
+    await updatePassword(oldPassword, password, user_id);
+
+    //const updateUserJwt = createToken(updateUserDB?.user_id);
+
+    response.status(200).json({ result: email, message: "User updated" });
   } catch (error) {
     console.log("error:", error);
     response.status(404).json({ error: error.message });
@@ -175,6 +313,18 @@ const updateUserAPI = async (request, response) => {
 };
 
 const deleteUser = async function (user_id) {
+  let query_1 = {
+    text: "select user_id from quatro_user where user_id=$1",
+    values: [user_id],
+  };
+
+  let resultQuery_1 = await pool.query(query_1);
+  let user = resultQuery_1.rows;
+
+  if (user.length === 0) {
+    throw Error("User doesn't exist");
+  }
+
   let query = {
     text: "delete from quatro_user where user_id = $1",
     values: [user_id],
@@ -189,7 +339,7 @@ const deleteUserAPI = async (request, response) => {
   const { user_id } = request.body;
   try {
     let userDelete = await deleteUser(user_id);
-    response.status(200).json({ result: userDelete });
+    response.status(200).json({ result: userDelete, message: "User deleted" });
   } catch (error) {
     console.log("error:", error);
     response.status(404).json({ error: error.message });
@@ -201,5 +351,6 @@ module.exports = {
   loginAPI,
   createUserAPI,
   updateUserAPI,
+  updatePasswordAPI,
   deleteUserAPI,
 };
